Integration of Customizer
I've deleted the theme Options page and instead switched the code to
use the theme customizer.

I also deleted the GPL license and instead link directly to the GNU
website as per theme review specs.

I tweak the CSS here and there. There was some stuff in there that
needed to be fixed everytime. It's very minor but should give you less
headaches in the future.