Added font size and blockquote p declaration
Since most text that gets highlighted as a blockquote in the editor of
WordPress has a p tag inside of it. the declaration of a font size is
over ruled by line 110. This resolves that issue.

Also since it is also a p tag it has a margin bottom to it and if you
are using a border top and border bottom the space between the quote
looks un even. This also resolves that issue.